refactor: Moved non-view functions to viewer.utils
