+++ viewer/utils.py
@@ -1,8 +1,3 @@
-"""
-utils.py
-
-Collection of technical methods tidied up in one location.
-"""
 import fnmatch
 import itertools
 import json
@@ -11,18 +6,25 @@ import os
 import shutil
 import string
 import tempfile
+import uuid
+import zipfile
 from pathlib import Path
 from typing import Dict, Generator, Optional
 from urllib.parse import urlparse
 
+import pandas as pd
 from django.conf import settings
 from django.contrib.auth.models import User
+from django.core.files.base import ContentFile
+from django.core.files.storage import default_storage
+from django.core.mail import send_mail
 from django.db import IntegrityError, transaction
 from django.db.models import F
 from django.http import JsonResponse
 from rdkit import Chem
 
 from scoring.models import SiteObservationGroup, SiteObvsSiteObservationGroup
+from viewer.views import logger
 
 from .models import (
     SiteObservation,
@@ -430,3 +432,135 @@ def alphanumerator(start_from: str = "") -> Generator[str, None, None]:
         _ = next(generator)
 
     return generator
+
+
+def save_pdb_zip(pdb_file):
+    zf = zipfile.ZipFile(pdb_file)
+    zip_lst = zf.namelist()
+    zfile = {}
+    zfile_hashvals: Dict[str, str] = {}
+    print(zip_lst)
+    for filename in zip_lst:
+        # only handle pdb files
+        if filename.split('.')[-1] == 'pdb':
+            f = filename.split('/')[0]
+            save_path = os.path.join(settings.MEDIA_ROOT, 'tmp', f)
+            if default_storage.exists(f):
+                rand_str = uuid.uuid4().hex
+                pdb_path = default_storage.save(
+                    save_path.replace('.pdb', f'-{rand_str}.pdb'),
+                    ContentFile(zf.read(filename)),
+                )
+            # Test if Protein object already exists
+            # code = filename.split('/')[-1].replace('.pdb', '')
+            # test_pdb_code = filename.split('/')[-1].replace('.pdb', '')
+            # test_prot_objs = Protein.objects.filter(code=test_pdb_code)
+            #
+            # if len(test_prot_objs) > 0:
+            #     # make a unique pdb code as not to overwrite existing object
+            #     rand_str = uuid.uuid4().hex
+            #     test_pdb_code = f'{code}#{rand_str}'
+            #     zfile_hashvals[code] = rand_str
+            #
+            # fn = test_pdb_code + '.pdb'
+            #
+            # pdb_path = default_storage.save('tmp/' + fn,
+            #                                 ContentFile(zf.read(filename)))
+            else:
+                pdb_path = default_storage.save(
+                    save_path, ContentFile(zf.read(filename))
+                )
+            test_pdb_code = pdb_path.split('/')[-1].replace('.pdb', '')
+            zfile[test_pdb_code] = pdb_path
+
+    # Close the zip file
+    if zf:
+        zf.close()
+
+    return zfile, zfile_hashvals
+
+
+def save_tmp_file(myfile):
+    """Save file in temporary location for validation/upload processing"""
+
+    name = myfile.name
+    path = default_storage.save('tmp/' + name, ContentFile(myfile.read()))
+    tmp_file = str(os.path.join(settings.MEDIA_ROOT, path))
+
+    return tmp_file
+
+
+def create_csv_from_dict(input_dict, title=None, filename=None):
+    """Write a CSV file containing data from an input dictionary and return a URL
+    to the file in the media directory.
+    """
+    if not filename:
+        filename = 'download'
+
+    media_root = settings.MEDIA_ROOT
+    unique_dir = str(uuid.uuid4())
+    # /code/media/downloads/unique_dir
+    download_path = os.path.join(media_root, 'downloads', unique_dir)
+    os.makedirs(download_path, exist_ok=True)
+
+    download_file = os.path.join(download_path, filename)
+
+    # Remove file if it already exists
+    if os.path.isfile(download_file):
+        os.remove(download_file)
+
+    with open(download_file, "w", newline='', encoding='utf-8') as csvfile:
+        if title:
+            csvfile.write(title)
+            csvfile.write("\n")
+
+    df = pd.DataFrame.from_dict(input_dict)
+    df.to_csv(download_file, mode='a', header=True, index=False)
+
+    return download_file
+
+
+def email_task_completion(
+    contact_email, message_type, target_name, target_path=None, task_id=None
+):
+    """Notify user of upload completion"""
+
+    logger.info('+ email_notify_task_completion: ' + message_type + ' ' + target_name)
+    email_from = settings.EMAIL_HOST_USER
+
+    if contact_email == '' or not email_from:
+        # Only send email if configured.
+        return
+
+    if message_type == 'upload-success':
+        subject = 'Fragalysis: Target: ' + target_name + ' Uploaded'
+        message = (
+            'The upload of your target data is complete. Your target is available at: '
+            + str(target_path)
+        )
+    elif message_type == 'validate-success':
+        subject = 'Fragalysis: Target: ' + target_name + ' Validation'
+        message = (
+            'Your data was validated. It can now be uploaded using the upload option.'
+        )
+    else:
+        # Validation failure
+        subject = 'Fragalysis: Target: ' + target_name + ' Validation/Upload Failed'
+        message = (
+            'The validation/upload of your target data did not complete successfully. '
+            'Please navigate the following link to check the errors: validate_task/'
+            + str(task_id)
+        )
+
+    recipient_list = [
+        contact_email,
+    ]
+    logger.info('+ email_notify_task_completion email_from: %s', email_from)
+    logger.info('+ email_notify_task_completion subject: %s', subject)
+    logger.info('+ email_notify_task_completion message: %s', message)
+    logger.info('+ email_notify_task_completion contact_email: %s', contact_email)
+
+    # Send email - this should not prevent returning to the screen in the case of error.
+    send_mail(subject, message, email_from, recipient_list, fail_silently=True)
+    logger.info('- email_notify_task_completion')
+    return
