+++ viewer/views.py
@@ -3,7 +3,6 @@ import logging
 import os
 import shlex
 import shutil
-import uuid
 import zipfile
 from datetime import datetime
 from pathlib import Path
@@ -18,7 +17,6 @@ from dateutil.parser import parse
 from django.conf import settings
 from django.core.files.base import ContentFile
 from django.core.files.storage import default_storage
-from django.core.mail import send_mail
 from django.db import connections
 from django.http import FileResponse, HttpResponse, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
@@ -45,7 +43,12 @@ from viewer.squonk2_agent import (
     Squonk2AgentRv,
     get_squonk2_agent,
 )
-from viewer.utils import create_squonk_job_request_url, handle_uploaded_file
+from viewer.utils import (
+    create_csv_from_dict,
+    create_squonk_job_request_url,
+    handle_uploaded_file,
+    save_tmp_file,
+)
 
 from .discourse import (
     check_discourse_user,
@@ -90,9 +93,12 @@ _SQ2A: Squonk2Agent = get_squonk2_agent()
 
 
 def react(request):
-    """We "START HERE". This is the first API call that the front-end calls."""
-
-    discourse_api_key = settings.DISCOURSE_API_KEY
+    """
+    The f/e starts here.
+    This is the first API call that the front-end calls, and it returns a 'context'
+    defining the state of things like the legacy URL, Squonk and Discourse availability
+    via the 'context' variable used for the view's template.
+    """
 
     # Start building the context that will be passed to the template
     context = {'legacy_url': settings.LEGACY_URL}
@@ -107,13 +113,14 @@ def react(request):
         logger.info("Squonk2 is NOT configured")
         context['squonk_available'] = 'false'
 
+    discourse_api_key = settings.DISCOURSE_API_KEY
     context['discourse_available'] = 'true' if discourse_api_key else 'false'
     user = request.user
     if user.is_authenticated:
         context['discourse_host'] = ''
         context['user_present_on_discourse'] = 'false'
-        # If user is authenticated and a discourse api key is available, then check discourse to
-        # see if user is set up and set up flag in context.
+        # If user is authenticated and a discourse api key is available,
+        # hen check discourse to see if user is set up and set up flag in context.
         if discourse_api_key:
             context['discourse_host'] = settings.DISCOURSE_HOST
             _, _, user_id = check_discourse_user(user)
@@ -122,144 +129,12 @@ def react(request):
 
         # User is authenticated, so if Squonk can be called
         # return the Squonk UI URL
-        # so the f/e knows where to go.
+        # so the f/e knows where to go to find it.
         context['squonk_ui_url'] = ''
         if sq2_rv.success and check_squonk_active(request):
             context['squonk_ui_url'] = _SQ2A.get_ui_url()
 
     return render(request, "viewer/react_temp.html", context)
-
-
-def save_pdb_zip(pdb_file):
-    zf = zipfile.ZipFile(pdb_file)
-    zip_lst = zf.namelist()
-    zfile = {}
-    zfile_hashvals: Dict[str, str] = {}
-    print(zip_lst)
-    for filename in zip_lst:
-        # only handle pdb files
-        if filename.split('.')[-1] == 'pdb':
-            f = filename.split('/')[0]
-            save_path = os.path.join(settings.MEDIA_ROOT, 'tmp', f)
-            if default_storage.exists(f):
-                rand_str = uuid.uuid4().hex
-                pdb_path = default_storage.save(
-                    save_path.replace('.pdb', f'-{rand_str}.pdb'),
-                    ContentFile(zf.read(filename)),
-                )
-            # Test if Protein object already exists
-            # code = filename.split('/')[-1].replace('.pdb', '')
-            # test_pdb_code = filename.split('/')[-1].replace('.pdb', '')
-            # test_prot_objs = Protein.objects.filter(code=test_pdb_code)
-            #
-            # if len(test_prot_objs) > 0:
-            #     # make a unique pdb code as not to overwrite existing object
-            #     rand_str = uuid.uuid4().hex
-            #     test_pdb_code = f'{code}#{rand_str}'
-            #     zfile_hashvals[code] = rand_str
-            #
-            # fn = test_pdb_code + '.pdb'
-            #
-            # pdb_path = default_storage.save('tmp/' + fn,
-            #                                 ContentFile(zf.read(filename)))
-            else:
-                pdb_path = default_storage.save(
-                    save_path, ContentFile(zf.read(filename))
-                )
-            test_pdb_code = pdb_path.split('/')[-1].replace('.pdb', '')
-            zfile[test_pdb_code] = pdb_path
-
-    # Close the zip file
-    if zf:
-        zf.close()
-
-    return zfile, zfile_hashvals
-
-
-def save_tmp_file(myfile):
-    """Save file in temporary location for validation/upload processing"""
-
-    name = myfile.name
-    path = default_storage.save('tmp/' + name, ContentFile(myfile.read()))
-    tmp_file = str(os.path.join(settings.MEDIA_ROOT, path))
-
-    return tmp_file
-
-
-def create_csv_from_dict(input_dict, title=None, filename=None):
-    """Write a CSV file containing data from an input dictionary and return a URL
-    to the file in the media directory.
-    """
-    if not filename:
-        filename = 'download'
-
-    media_root = settings.MEDIA_ROOT
-    unique_dir = str(uuid.uuid4())
-    # /code/media/downloads/unique_dir
-    download_path = os.path.join(media_root, 'downloads', unique_dir)
-    os.makedirs(download_path, exist_ok=True)
-
-    download_file = os.path.join(download_path, filename)
-
-    # Remove file if it already exists
-    if os.path.isfile(download_file):
-        os.remove(download_file)
-
-    with open(download_file, "w", newline='', encoding='utf-8') as csvfile:
-        if title:
-            csvfile.write(title)
-            csvfile.write("\n")
-
-    df = pd.DataFrame.from_dict(input_dict)
-    df.to_csv(download_file, mode='a', header=True, index=False)
-
-    return download_file
-
-
-def email_task_completion(
-    contact_email, message_type, target_name, target_path=None, task_id=None
-):
-    """Notify user of upload completion"""
-
-    logger.info('+ email_notify_task_completion: ' + message_type + ' ' + target_name)
-    email_from = settings.EMAIL_HOST_USER
-
-    if contact_email == '' or not email_from:
-        # Only send email if configured.
-        return
-
-    if message_type == 'upload-success':
-        subject = 'Fragalysis: Target: ' + target_name + ' Uploaded'
-        message = (
-            'The upload of your target data is complete. Your target is available at: '
-            + str(target_path)
-        )
-    elif message_type == 'validate-success':
-        subject = 'Fragalysis: Target: ' + target_name + ' Validation'
-        message = (
-            'Your data was validated. It can now be uploaded using the upload option.'
-        )
-    else:
-        # Validation failure
-        subject = 'Fragalysis: Target: ' + target_name + ' Validation/Upload Failed'
-        message = (
-            'The validation/upload of your target data did not complete successfully. '
-            'Please navigate the following link to check the errors: validate_task/'
-            + str(task_id)
-        )
-
-    recipient_list = [
-        contact_email,
-    ]
-    logger.info('+ email_notify_task_completion email_from: %s', email_from)
-    logger.info('+ email_notify_task_completion subject: %s', subject)
-    logger.info('+ email_notify_task_completion message: %s', message)
-    logger.info('+ email_notify_task_completion contact_email: %s', contact_email)
-
-    # Send email - this should not prevent returning to the screen in the case of error.
-    send_mail(subject, message, email_from, recipient_list, fail_silently=True)
-    logger.info('- email_notify_task_completion')
-    return
 
 
 # --------------------
